Refactor Conversion Passes
This is a combination of 3 commits.

save changes

fix bug

 remove NVVM and ROCDL from TritonToTritonGPU pass
