+++ .gitignore
@@ -18,3 +18,6 @@ __pycache__
 # JetBrains project files
 .idea
 cmake-build-*
+
+log_*
+scripts*
