+++ bin/triton-opt.cpp
@@ -4,7 +4,8 @@
 #include "triton/Dialect/Triton/Transforms/Passes.h"
 #include "triton/Dialect/TritonGPU/Transforms/Passes.h"
 
-#include "triton/Conversion/Passes.h"
+#include "triton/Conversion/TritonToTritonGPU/Passes.h"
+#include "triton/Conversion/TritonGPUToLLVM/Passes.h"
 
 #include "mlir/IR/Dialect.h"
 #include "mlir/InitAllPasses.h"
