+++ include/triton/Conversion/CMakeLists.txt
@@ -1,3 +1,2 @@
-set(LLVM_TARGET_DEFINITIONS Passes.td)
-mlir_tablegen(Passes.h.inc -gen-pass-decls)
-add_public_tablegen_target(TritonConversionPassIncGen)
+add_subdirectory(TritonToTritonGPU)
+add_subdirectory(TritonGPUToLLVM)
