+++ include/triton/Conversion/TritonGPUToLLVM/Passes.td
@@ -1,28 +1,7 @@
-#ifndef TRITON_CONVERSION_PASSES
-#define TRITON_CONVERSION_PASSES
+#ifndef TRITONGPU_CONVERSION_PASSES
+#define TRITONGPU_CONVERSION_PASSES
 
 include "mlir/Pass/PassBase.td"
-
-def ConvertTritonToTritonGPU: Pass<"convert-triton-to-tritongpu", "mlir::ModuleOp"> {
-    let summary = "Convert Triton to TritonGPU";
-    let description = [{
-
-    }];
-    let constructor = "mlir::triton::createConvertTritonToTritonGPUPass()";
-
-    let dependentDialects = ["mlir::arith::ArithDialect",
-                             "mlir::math::MathDialect",
-                             // TODO: Does this pass depend on SCF?
-                             "mlir::scf::SCFDialect",
-                             "mlir::triton::TritonDialect",
-                             "mlir::triton::gpu::TritonGPUDialect"];
-
-   let options = [
-       Option<"numWarps", "num-warps",
-              "int32_t", /*default*/"4",
-              "number of warps">
-   ];
-}
 
 
 def ConvertTritonGPUToLLVM : Pass<"convert-triton-gpu-to-llvm", "mlir::ModuleOp"> {
