+++ include/triton/Conversion/TritonToTritonGPU/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(LLVM_TARGET_DEFINITIONS Passes.td)
+mlir_tablegen(Passes.h.inc -gen-pass-decls --name TritonToTritonGPU)
+add_public_tablegen_target(TritonConversionPassIncGen)
