+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -19,9 +19,10 @@ add_mlir_conversion_library(TritonGPUToLLVM
 
     ADDITIONAL_HEADER_DIRS
     ${PROJECT_SOURCE_DIR}/include/triton/Conversion/TritonGPUToLLVM
+    ${PROJECT_BINARY_DIR}/include/triton/Conversion/TritonGPUToLLVM
 
     DEPENDS
-    TritonConversionPassIncGen
+    TritonGPUConversionPassIncGen
 
     LINK_COMPONENTS
     Core
