+++ lib/Conversion/TritonToTritonGPU/CMakeLists.txt
@@ -3,6 +3,7 @@ add_mlir_conversion_library(TritonToTritonGPU
 
     ADDITIONAL_HEADER_DIRS
     ${PROJECT_SOURCE_DIR}/include/triton/Conversion/TritonToTritonGPU
+    ${PROJECT_BINARY_DIR}/include/triton/Conversion/TritonToTritonGPU
 
     DEPENDS
     TritonConversionPassIncGen
